change(api): move some api types to different modules to allow for more custom loading of deep agent
